Add is_reverse function to string_chap8.py
This function checks if one string is the reverse of another by comparing characters from the start of the first string to the end of the second string. It returns False if the strings have different lengths or if any characters do not match. Additionally, an assertion is included to test the function.