Add age attribute to Person and prioritize queue by age
The `Person` class now includes an `age` attribute, which is used to modify queue behavior in the `Shop` class. People aged 50 or older are added to the end of the queue, while others are added to the front. Updated the `__str__` and `__repr__` methods and adjusted the main function to reflect these changes.